+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - sdk(sdk): Added deposit simulation.
 - sdk(sdk): Added withdrawal simulation.
 - sdk(sdk): Added GLV deposit simulation.
+- sdk(sdk): Added GLV withdrawal simulation.
 - cli: Added auto-creation support of multiple ALTs based on ALT limit for `alt extend` command.
 
 ### Changed